backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.3
websockets==14.1
python-multipart==0.0.20

# Whisper API integration
openai>=1.0.0

# AI features (optional)
google-generativeai==0.8.3
python-dotenv==1.2.1
httpx==0.28.1
anthropic>=0.18.0
faster-whisper==0.10.0
